Added support for Barney Mouse and Wizard's Castle.  Renamed sextettiii to soft_touch and moved platou into soft_touch
