+++ libdisk/include/libdisk/track_types.h
@@ -335,3 +335,6 @@ X(turn_it, "Turn It")
 X(platou, "Platou")
 X(evolution_cryser, "Evolution Cryser")
 X(mind_force, "Mind Force")
+X(barney_mouse, "Barney Mouse")
+X(wizards_castle_a, "Wizard's Castle A")
+X(wizards_castle_b, "Wizard's Castle B")
